requirements.txt
# Core web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Image processing
numpy>=2.0.0
Pillow>=10.0.0
opencv-python-headless>=4.9.0

# HTTP client for Replicate API
httpx>=0.27.0

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0

# File uploads
python-multipart>=0.0.7

# Async file operations
aiofiles>=23.0.0
